Refactor: 모달 컴포넌트 MVVM 패턴 적용 #39
