+++ src/components/default/modal/ModalView.tsx
@@ -1,249 +1,59 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import Modal from 'react-modal';
-import { useForm } from 'react-hook-form';
-import { useMutation, useQueryClient } from 'react-query';
-
-import SelectArrowIcon from 'assets/main/main_modal_select_arrow.svg';
 import * as S from './ModalStyledComponents';
-import { processTypeInfo, processTypeList } from 'constants/process';
-import { IRegisterNewApplication } from 'types/interfaces/KanbanProcess';
 import { formatProcessToKor } from 'utils/process';
-import {
-  registerSimpleApplication,
-  editSimpleApplication,
-  updateApplicationProcess,
-  createNewProcess,
-} from 'apis/kanban';
+import { processTypeInfo, processTypeList } from 'constants/process';
+import SelectArrowIcon from 'assets/main/main_modal_select_arrow.svg';
+import { FieldValues, UseFormGetValues } from 'react-hook-form';
 
-interface IProps {
-  mode: string;
-  modalIsOpen: boolean;
-  closeModal: () => void;
-  currentProcessType: string;
-  fetchedProcessData: any;
-  applicationInfo: any;
+interface ModalViewModelProps {
+  register: any; // 적절한 타입으로 변경해주세요.
+  handleSubmit: any; // 적절한 타입으로 변경해주세요.
+  errors: any; // 적절한 타입으로 변경해주세요.
+  dropDownToggle: boolean;
+  detailDropDownToggle: boolean;
+  dropDownToggleHandler: (dropDownId: string) => void;
+  dropDownItemHandler: (dropDownId: string, process: string) => void; // 적절한 타입으로 변경해주세요.
+  userInputToggle: boolean;
+  setUserInputToggle: React.Dispatch<React.SetStateAction<boolean>>;
+  handleApplicationSubmission: () => void;
+  isDetailedProcessTypeRequired: () => boolean;
+  currentProcessType: any; // 적절한 타입으로 변경해주세요.
+  fetchedProcessData: any; // 적절한 타입으로 변경해주세요.
+  applicationInfo: any; // 적절한 타입으로 변경해주세요.
+  mode: any; // 적절한 타입으로 변경해주세요.
+  getValues: UseFormGetValues<FieldValues>;
+  defaultValues: any; // 적절한 타입으로 변경해주세요.
+  validationProcess: () => void;
 }
 
-/*
-TODO
-  1. as 타입 추론 제거
-  2. any 타입 제거
-  3. 옵셔널 체이닝 제거
-*/
-const ModalComponent = ({
-  mode,
-  modalIsOpen,
-  closeModal,
-  currentProcessType,
-  fetchedProcessData,
-  applicationInfo,
-}: IProps) => {
-  const queryClient = useQueryClient();
-  const [dropDownToggle, setDropDownToggle] = useState(false);
-  const [detailDropDownToggle, setDetailDropDownToggle] = useState(false);
-  const [userInputToggle, setUserInputToggle] = useState(false);
+interface IProps {
+  viewModel: ModalViewModelProps;
+  modalIsOpen: boolean;
+  closeModal: () => void;
+}
 
+const ModalView = ({ viewModel, modalIsOpen, closeModal }: IProps) => {
   const {
     register,
     handleSubmit,
+    errors,
+    dropDownToggle,
+    detailDropDownToggle,
+    dropDownToggleHandler,
+    dropDownItemHandler,
+    userInputToggle,
+    setUserInputToggle,
+    handleApplicationSubmission,
+    isDetailedProcessTypeRequired,
+    currentProcessType,
+    fetchedProcessData,
+    applicationInfo,
+    mode,
     getValues,
-    setValue,
-    reset,
-    setError,
-    clearErrors,
-    watch,
-    formState: { errors, defaultValues },
-  } = useForm({
-    mode: 'onSubmit',
-  });
-
-  const detailedProcessType = watch('detailedProcessType');
-
-  function invalidateKanbanListOnSuccess() {
-    queryClient.invalidateQueries('fetchKanbanList');
-  }
-
-  const registerMutation = useMutation(
-    (newApplicationData: IRegisterNewApplication) => registerSimpleApplication(newApplicationData),
-    {
-      onSuccess() {
-        invalidateKanbanListOnSuccess();
-      },
-    },
-  );
-
-  const editMutation = useMutation(
-    (editData: any) => {
-      const { editApplicationData, applicationId } = editData;
-      return editSimpleApplication(editApplicationData, applicationId);
-    },
-    {
-      onSuccess() {
-        invalidateKanbanListOnSuccess();
-      },
-    },
-  );
-
-  const updateProcessMutation = useMutation(
-    (updateData: any) => {
-      const { applicationId, processId } = updateData;
-      return updateApplicationProcess(applicationId, processId);
-    },
-    {
-      onSuccess() {
-        invalidateKanbanListOnSuccess();
-        closeModal();
-      },
-    },
-  );
-
-  function isDetailedProcessTypeRequired() {
-    if (getValues('processType') === 'TO_APPLY' || getValues('processType') === 'DOCUMENT') {
-      return false;
-    } else {
-      return true;
-    }
-  }
-
-  // 전형 or 세부단계 드롭박스 토글 함수
-  function dropDownToggleHandler(dropDownId: string) {
-    if (dropDownId === 'processType') {
-      setDropDownToggle(!dropDownToggle);
-      setDetailDropDownToggle(false);
-    } else {
-      setDetailDropDownToggle(!detailDropDownToggle);
-    }
-  }
-
-  // 세부단계 유효성 검사 함수
-  function validationProcess() {
-    if (isDetailedProcessTypeRequired()) {
-      if (!detailedProcessType) {
-        setError('detailedProcessType', {
-          type: 'required',
-          message: '세부 단계를 선택하세요',
-        });
-      } else {
-        clearErrors('detailedProcessType');
-      }
-    }
-  }
-
-  // 전형 or 세부 단계 선택시 실행 함수
-  function dropDownItemHandler(dropDownId: string, process: string) {
-    if (dropDownId === 'processType') {
-      setValue('processType', process);
-      setValue('detailedProcessType', '');
-      setDropDownToggle(!dropDownToggle);
-    } else {
-      setValue('detailedProcessType', process);
-      setDetailDropDownToggle(!detailDropDownToggle);
-    }
-  }
-
-  async function handleApplicationSubmission() {
-    if (mode === 'simpleRegister') {
-      const newApplicationData = {
-        companyName: getValues('companyName'),
-        position: getValues('position'),
-        url: getValues('url'),
-        currentProcess: {
-          type: getValues('processType'),
-          description:
-            getValues('detailedProcessType') === ''
-              ? formatProcessToKor(getValues('processType'))
-              : getValues('detailedProcessType'),
-          schedule: getValues('schedule'),
-        },
-      };
-
-      registerMutation.mutate(newApplicationData);
-    } else if (mode === 'simpleEdit') {
-      const editApplicationData = {
-        companyName: getValues('companyName'),
-        position: getValues('position'),
-        processId: applicationInfo.process.id,
-        schedule: getValues('schedule'),
-        url: getValues('url'),
-      };
-
-      editMutation.mutate({
-        editApplicationData,
-        applicationId: applicationInfo.applicationId,
-      });
-    } else if (mode === 'updateCurrentProcess') {
-      const newProcessData = {
-        type: getValues('processType'),
-        description:
-          getValues('detailedProcessType') === ''
-            ? getValues('processType')
-            : getValues('detailedProcessType'),
-      };
-
-      const data = await createNewProcess(applicationInfo.applicationId, newProcessData);
-
-      if (data.success) {
-        updateProcessMutation.mutate({
-          applicationId: applicationInfo.applicationId,
-          processId: data.data.id,
-        });
-      }
-    }
-    closeModal();
-  }
-
-  function getResetValues() {
-    if (!modalIsOpen) {
-      return {
-        companyName: '',
-        processType: '',
-        detailedProcessType: '',
-        position: '',
-        schedule: '',
-        url: '',
-      };
-    }
-
-    const processType = currentProcessType ? currentProcessType : applicationInfo.process.type;
-
-    const detailedProcessType =
-      mode === 'updateCurrentProcess'
-        ? applicationInfo.processDescription.length > 0
-          ? applicationInfo.processDescription[0].description
-          : ''
-        : applicationInfo.process.description;
-
-    return {
-      companyName: applicationInfo.companyName,
-      processType,
-      detailedProcessType,
-      position: applicationInfo.position,
-      schedule: applicationInfo.process.schedule,
-      url: applicationInfo.url,
-    };
-  }
-
-  function resetInputValues() {
-    setValue('detailedProcessType', '');
-  }
-
-  useEffect(() => {
-    if (userInputToggle) {
-      resetInputValues();
-    }
-  }, [userInputToggle]);
-
-  useEffect(() => {
-    reset(getResetValues());
-
-    setDropDownToggle(false);
-    setDetailDropDownToggle(false);
-    setUserInputToggle(false);
-  }, [modalIsOpen, mode]);
-
-  useEffect(() => {
-    validationProcess();
-  }, [getValues('detailedProcessType')]);
+    defaultValues,
+    validationProcess,
+  } = viewModel;
 
   if (mode === 'updateCurrentProcess') {
     return (
@@ -506,4 +316,4 @@ const ModalComponent = ({
     );
 };
 
-export default ModalComponent;
+export default ModalView;
